Switch to fully Composable UI
