+++ app/src/main/java/com/boswelja/contactringtonegenerator/ringtonebuilder/RingtoneCreatorScreen.kt
@@ -1,13 +1,6 @@
 package com.boswelja.contactringtonegenerator.ringtonebuilder
 
-import android.media.RingtoneManager
 import android.net.Uri
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -27,20 +20,16 @@ import androidx.compose.material.DismissDirection
 import androidx.compose.material.DismissValue
 import androidx.compose.material.Divider
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.ExtendedFloatingActionButton
-import androidx.compose.material.FabPosition
 import androidx.compose.material.Icon
 import androidx.compose.material.ListItem
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.Scaffold
 import androidx.compose.material.Surface
 import androidx.compose.material.SwipeToDismiss
 import androidx.compose.material.Text
 import androidx.compose.material.TextButton
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Delete
-import androidx.compose.material.icons.outlined.NavigateNext
 import androidx.compose.material.rememberDismissState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -49,77 +38,14 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.boswelja.contactringtonegenerator.R
-import com.boswelja.contactringtonegenerator.WizardViewModel
-import com.boswelja.contactringtonegenerator.common.ui.AppTheme
 import com.boswelja.contactringtonegenerator.ringtonegen.item.StructureChoice
 import com.boswelja.contactringtonegenerator.ringtonegen.item.StructureItem
 import com.boswelja.contactringtonegenerator.ringtonegen.item.Utils
 import timber.log.Timber
-
-class RingtoneCreatorFragment : Fragment() {
-
-    private val wizardModel: WizardViewModel by activityViewModels()
-
-    private val audioPickerLauncher: ActivityResultLauncher<String> = registerForActivityResult(
-        ActivityResultContracts.GetContent()
-    ) { }
-
-    private val ringtonePickerLauncher: ActivityResultLauncher<Int> = registerForActivityResult(
-        PickRingtone()
-    ) { }
-
-    @ExperimentalMaterialApi
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                val viewModel: RingtoneCreatorViewModel = viewModel()
-                val structure = viewModel.ringtoneStructure
-                AppTheme {
-                    Scaffold(
-                        floatingActionButtonPosition = FabPosition.End,
-                        floatingActionButton = {
-                            ExtendedFloatingActionButton(
-                                text = { Text(stringResource(R.string.next)) },
-                                icon = { Icon(Icons.Outlined.NavigateNext, null) },
-                                onClick = {
-                                    wizardModel.submitRingtoneStructure(structure)
-                                }
-                            )
-                        }
-                    ) {
-                        RingtoneBuilderScreen(
-                            structure = structure,
-                            onItemAdded = { item -> viewModel.ringtoneStructure.add(item) },
-                            onActionClicked = { item ->
-                                when (item.dataType) {
-                                    StructureItem.DataType.AUDIO_FILE ->
-                                        audioPickerLauncher.launch("audio/*")
-                                    StructureItem.DataType.SYSTEM_RINGTONE ->
-                                        ringtonePickerLauncher.launch(RingtoneManager.TYPE_RINGTONE)
-                                    else -> Timber.w("Unknown action clicked")
-                                }
-                            },
-                            onItemRemoved = { item -> viewModel.ringtoneStructure.remove(item) },
-                            modifier = Modifier.padding(it)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
 
 @ExperimentalMaterialApi
 @Composable
