+++ app/src/main/java/com/boswelja/contactringtonegenerator/settings/ui/SettingsScreen.kt
@@ -1,10 +1,5 @@
 package com.boswelja.contactringtonegenerator.settings.ui
 
-import android.content.Intent
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.ExperimentalMaterialApi
@@ -12,47 +7,12 @@ import androidx.compose.material.ListItem
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.stringResource
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.boswelja.contactringtonegenerator.R
-import com.boswelja.contactringtonegenerator.common.ui.AppTheme
 import com.boswelja.contactringtonegenerator.ui.common.CheckboxPreference
 import com.boswelja.contactringtonegenerator.ui.common.SliderPreference
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-
-class SettingsFragment : Fragment() {
-
-    @ExperimentalCoroutinesApi
-    @ExperimentalMaterialApi
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setContent {
-                AppTheme {
-                    SettingsScreen(
-                        onLaunchTtsClick = {
-                            launchTtsSettings()
-                        }
-                    )
-                }
-            }
-        }
-    }
-
-    private fun launchTtsSettings() {
-        Intent().apply {
-            action = "com.android.settings.TTS_SETTINGS"
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK
-        }.also {
-            startActivity(it)
-        }
-    }
-}
 
 @ExperimentalCoroutinesApi
 @ExperimentalMaterialApi
