--- app/src/main/res/navigation/main_navigation.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/main_navigation"
-    app:startDestination="@id/getStartedFragment">
-
-    <fragment
-        android:id="@+id/getStartedFragment"
-        android:name="com.boswelja.contactringtonegenerator.entry.ui.GetStartedFragment">
-        <action
-            android:id="@+id/to_contactPickerFragment"
-            app:destination="@id/contactPickerFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:exitAnim="@anim/slide_out_left"
-            app:popExitAnim="@anim/slide_out_right" />
-        <action
-            android:id="@+id/to_settingsFragment"
-            app:destination="@id/settingsFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:exitAnim="@anim/slide_out_left"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-    <fragment
-        android:id="@+id/contactPickerFragment"
-        android:name="com.boswelja.contactringtonegenerator.contactpicker.ui.ContactPickerFragment"
-        android:label="@string/contact_picker_title">
-        <action
-            android:id="@+id/to_ringtoneCreatorFragment"
-            app:destination="@id/ringtoneCreatorFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:exitAnim="@anim/slide_out_left"
-            app:popExitAnim="@anim/slide_out_right" />
-    </fragment>
-    <fragment
-        android:id="@+id/ringtoneCreatorFragment"
-        android:label="Build a Ringtone"
-        android:name="com.boswelja.contactringtonegenerator.ringtonebuilder.RingtoneCreatorFragment">
-        <action
-            android:id="@+id/to_loadingFragment"
-            app:destination="@id/loadingFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right"
-            app:popUpTo="@id/getStartedFragment"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <fragment
-        android:id="@+id/loadingFragment"
-        android:name="com.boswelja.contactringtonegenerator.common.ui.ProgressFragment">
-        <action
-            android:id="@+id/to_finishedFragment"
-            app:destination="@id/resultsFragment"
-            app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_left"
-            app:popEnterAnim="@anim/slide_in_left"
-            app:popExitAnim="@anim/slide_out_right"
-            app:popUpTo="@id/loadingFragment"
-            app:popUpToInclusive="true" />
-    </fragment>
-    <fragment
-        android:id="@+id/settingsFragment"
-        android:name="com.boswelja.contactringtonegenerator.settings.ui.SettingsFragment"
-        android:label="@string/settings_title" />
-    <fragment
-        android:id="@+id/resultsFragment"
-        android:name="com.boswelja.contactringtonegenerator.result.ui.ResultsFragment">
-        <argument
-            android:name="result"
-            app:argType="com.boswelja.contactringtonegenerator.ringtonegen.Result" />
-    </fragment>
-</navigation>
