+++ app/src/main/res/values/styles.xml
@@ -1,27 +1,14 @@
 <resources>
 
-    <style name="BaseTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="BaseTheme" parent="Theme.AppCompat.DayNight.NoActionBar">
         <item name="colorPrimary">@color/primaryColor</item>
         <item name="colorPrimaryDark">@color/primaryDarkColor</item>
-        <item name="colorPrimaryVariant">@color/primaryLightColor</item>
-        <item name="colorOnPrimary">@color/primaryTextColor</item>
-        <item name="colorSecondary">@color/secondaryColor</item>
-        <item name="colorSecondaryVariant">@color/secondaryDarkColor</item>
-        <item name="colorOnSecondary">@color/secondaryTextColor</item>
         <item name="colorAccent">@color/primaryColor</item>
         <item name="android:windowBackground">@color/windowBackground</item>
         <item name="android:windowLightStatusBar">@bool/lightSystemBars</item>
         <item name="android:statusBarColor">@color/windowBackground</item>
-        <item name="textInputStyle">@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox.Dense</item>
     </style>
 
     <style name="AppTheme" parent="BaseTheme" />
-
-    <style name="WizardFabContinue">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_marginBottom">@dimen/content_side_margins</item>
-        <item name="android:layout_marginEnd">@dimen/content_side_margins</item>
-    </style>
 
 </resources>
